Fix broken if clause causing wrong ban times
Also remove the negative part of the randomized ban duration thing. What if a player is unbanned 3 days prior to his birthday and fails the test again? We now go up to 5 days into the future, that should be enough by far.
The if clause was true even when the age entered and the age from the birthday resulted in the same overall age, because ageverify_getAge returns an exact float rather than a floored int. We fix that, too.

We'll see if that fixed these weird negative-time bans.